data-ingestion/requirements.txt
python-dotenv
openai
pinecone
tiktoken
langchain 